split bumpover test modules
